+++ packages/core/popup/examples/11-asynchronous-popup.tsx
@@ -2,7 +2,6 @@
 import { FC, useState, useEffect } from 'react';
 import Button from '@atlaskit/button';
 import { Placement } from '@atlaskit/popper';
-import Spinner from '@atlaskit/spinner';
 import { jsx } from '@emotion/core';
 
 import Popup from '../src';
@@ -28,10 +27,6 @@ const contentCSS = {
   maxWidth: '300px',
 } as const;
 
-const spinnerContainerCSS = {
-  margin: '40px',
-};
-
 const expanderCSS = ({ width }: { width: number }) => ({
   display: 'inline-block',
   width: width ? `${width}px` : 0,
@@ -43,22 +38,27 @@ const PopupContent: FC<PopupProps> = ({
   position,
   setButtonWidth,
   buttonWidth,
+  scheduleUpdate,
 }) => {
   const [content, setContent] = useState(
     'Lorem Ipsum dolor sit amet. Lorem Ipsum dolor sit amet. Lorem Ipsum dolor sit amet. ',
   );
   const addContent = () => {
     setContent(`${content}Lorem Ipsum dolor sit amet. `);
+
+    // Reposition the popup
+    typeof scheduleUpdate === 'function' && scheduleUpdate();
   };
 
   const clearContent = () => {
     setContent('');
+
+    // Reposition the popup
+    typeof scheduleUpdate === 'function' && scheduleUpdate();
   };
 
   return loading ? (
-    <div id="spinner" css={spinnerContainerCSS}>
-      <Spinner size="large" />
-    </div>
+    <div id="spinner">Loading...</div>
   ) : (
     <div id="popup-content" css={contentCSS}>
       <Button onClick={() => setPosition()}>Toggle Position</Button>
@@ -127,13 +127,14 @@ export default () => {
       <Popup
         isOpen={isOpen}
         onClose={() => setIsOpen(false)}
-        content={() => (
+        content={({ scheduleUpdate }) => (
           <PopupContent
             loading={!isLoaded}
             setPosition={setPosition}
             position={position}
             setButtonWidth={setButtonWidth}
             buttonWidth={buttonWidth}
+            scheduleUpdate={scheduleUpdate}
           />
         )}
         trigger={triggerProps => (