+++ packages/editor/adf-schema/src/schema/nodes/index.ts
@@ -114,7 +114,13 @@ export {
   decisionListSelector,
 } from './decision-list';
 export { decisionItem, DecisionItemDefinition } from './decision-item';
-export { taskList, TaskListDefinition, taskListSelector } from './task-list';
+export {
+  taskList,
+  nestableTaskList,
+  TaskListDefinition,
+  TaskListWithNestingDefinition,
+  taskListSelector,
+} from './task-list';
 export { taskItem, TaskItemDefinition } from './task-item';
 export { extension, ExtensionDefinition } from './extension';
 export { inlineExtension, InlineExtensionDefinition } from './inline-extension';