+++ packages/editor/renderer/src/react/index.tsx
@@ -24,7 +24,7 @@ import {
   ExtensionHandlers,
   calcTableColumnWidths,
 } from '@atlaskit/editor-common';
-import { generateIdFromString } from '../utils';
+import { getText } from '../utils';
 
 export interface RendererContext {
   objectAri?: string;
@@ -45,6 +45,7 @@ export interface ConstructorParams {
   allowHeadingAnchorLinks?: boolean;
   allowColumnSorting?: boolean;
   fireAnalyticsEvent?: (event: AnalyticsEventPayload) => void;
+  shouldOpenMediaViewer?: boolean;
 }
 
 type MarkWithContent = Partial<Mark<any>> & {
@@ -89,6 +90,7 @@ export default class ReactSerializer implements Serializer<JSX.Element> {
   private allowHeadingAnchorLinks?: boolean;
   private allowColumnSorting?: boolean;
   private fireAnalyticsEvent?: (event: AnalyticsEventPayload) => void;
+  private shouldOpenMediaViewer?: boolean;
 
   constructor({
     providers,
@@ -102,6 +104,7 @@ export default class ReactSerializer implements Serializer<JSX.Element> {
     allowHeadingAnchorLinks,
     allowColumnSorting,
     fireAnalyticsEvent,
+    shouldOpenMediaViewer,
   }: ConstructorParams) {
     this.providers = providers;
     this.eventHandlers = eventHandlers;
@@ -114,6 +117,7 @@ export default class ReactSerializer implements Serializer<JSX.Element> {
     this.allowHeadingAnchorLinks = allowHeadingAnchorLinks;
     this.allowColumnSorting = allowColumnSorting;
     this.fireAnalyticsEvent = fireAnalyticsEvent;
+    this.shouldOpenMediaViewer = shouldOpenMediaViewer;
   }
 
   private resetState() {
@@ -148,19 +152,20 @@ export default class ReactSerializer implements Serializer<JSX.Element> {
           props = this.getHeadingProps(node, parentInfo && parentInfo.path);
         } else if (['tableHeader', 'tableRow'].indexOf(node.type.name) > -1) {
           props = this.getTableChildrenProps(node);
+        } else if (node.type.name === 'media') {
+          props = this.getMediaProps(node);
         } else {
           props = this.getProps(node);
         }
 
         let currentPath = (parentInfo && parentInfo.path) || [];
-        currentPath.push(node);
 
         const parentIsIncompleteTask =
           node.type.name === 'taskItem' && node.attrs.state !== 'DONE';
 
         let pInfo = {
           parentIsIncompleteTask,
-          path: currentPath,
+          path: [...currentPath, node],
         };
 
         const serializedContent = this.serializeFragment(
@@ -265,6 +270,12 @@ export default class ReactSerializer implements Serializer<JSX.Element> {
       parentIsIncompleteTask: parentInfo && parentInfo.parentIsIncompleteTask,
     };
   }
+  private getMediaProps(node: Node) {
+    return {
+      ...this.getProps(node),
+      shouldOpenMediaViewer: this.shouldOpenMediaViewer,
+    };
+  }
 
   private getProps(node: Node) {
     return {
@@ -302,12 +313,28 @@ export default class ReactSerializer implements Serializer<JSX.Element> {
     };
   }
 
+  // The return value of this function is NOT url encoded,
+  // In HTML5 standard, id can contain any characters, encoding is no necessary.
+  // Plus we trying to avoid double encoding, therefore we leave the value as is.
+  // Remember to use encodeURIComponent when generating url from the id value.
   private getHeadingId(node: Node) {
     if (this.disableHeadingIDs || !node.content.size) {
       return;
     }
 
-    return this.getUniqueHeadingId(generateIdFromString(node.textContent));
+    // We are not use node.textContent here, because we would like to handle cases where
+    // headings only contain inline blocks like emoji, status and date.
+    const nodeContent = (node as any).content
+      .toJSON()
+      .reduce((acc: string, node: any) => acc.concat(getText(node) || ''), '')
+      .trim()
+      .replace(/\s/g, '-');
+
+    if (!nodeContent) {
+      return;
+    }
+
+    return this.getUniqueHeadingId(nodeContent);
   }
 
   private getUniqueHeadingId(baseId: string, counter = 0): string {