+++ packages/people-and-teams/profilecard/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/profilecard
 
+## 12.3.3
+
+- Updated dependencies [bd94b1d552](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/bd94b1d552):
+- Updated dependencies [ae4f336a3a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ae4f336a3a):
+  - @atlaskit/util-data-test@13.0.0
+  - @atlaskit/i18n-tools@0.6.0
+
 ## 12.3.2
 
 ### Patch Changes